Add decrement button, use BLoC pattern
BLoC architecture, a (Business Logic Component) pattern, is a way to manage state (recommended by Google developers). It allows for separation of the business logic from UI by utilizing streams. Because business logic events are expressed though stream operators, we are able to use RxDart operations like 'distinct' and 'asyncMap' as we as 'withLatestFrom' to further mutate and filter that state.